+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Improvements:
 
 Bug fixes that are backwards-compatible:
 * [OLMIS-6317](https://openlmis.atlassian.net/browse/OLMIS-6317): Service account tokens will not expire
+* [OLMIS-6548](https://openlmis.atlassian.net/browse/OLMIS-6548): Add translations for reset password error modal.
 
 4.1.2 / 2019-05-27
 ==================